Updated dimensions to be of type size_t so we can allocate arrays larger than 2Gb.
git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1336 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/Misc/array.h
@@ -45,19 +45,19 @@
 #define aerr_set(e) ((e)?xerr_set((e), ArrayErrorString((e))):0)
 
 typedef struct {
-    int size;			/* element size */
-    int dim;			/* allocated number of elements */
-    int max;			/* elements accessed */
+    size_t size;			/* element size */
+    size_t dim;			/* allocated number of elements */
+    size_t max;			/* elements accessed */
     void *base;			/* base address of array */
 } ArrayStruct, *Array;
 
 
 
-extern Array ArrayCreate(size_t size, int dim);
+extern Array ArrayCreate(size_t size, size_t dim);
 
-extern int ArrayExtend(Array a, int dim);
+extern int ArrayExtend(Array a, size_t dim);
 
-extern void *ArrayRef(Array a, int i);
+extern void *ArrayRef(Array a, size_t i);
 
 extern int ArrayDestroy(Array a);
